fix transpose for square data
Problem was that `sort_idx` and `sort_N` attributes do not have the same length as `xvars` if some of the xvars have the same length as others (since `xvars` of the same length use the same `sort_idx` indices for shuffling, and `sort_idx` does not store duplicates).

The fix here is just to not let transpose try to swap the order of the `sort_idx` lists -- this is fine, since every time `sort_idx` is actually accessed to retrieve sort indices, it is done by finding the index of the element which has the same length as the xvar being shuffled/unshuffled. That is, the order of the elements in `sort_idx` (or `sort_N`) never mattered anyway.